==> notebooks/NB02-Build-Database.ipynb <==
**NB01-Data-Consoldation**

#!pip install pandas numpy lets-plot numerize tqdm sqlalchemy jupysql
%load_ext sql
import requests
from bs4 import BeautifulSoup
import pandas as pd
import uuid
import sqlite3

**Reading data from csv file from the folder /data/raw**

top_mov_actors_df = pd.read_csv('../data/raw/top_mov_actors.csv', index_col=0)
top_movies_df = pd.read_csv('../data/raw/top_movie.csv',index_col=0)
actors_df = pd.read_csv('../data/raw/actors_list.csv',index_col=0)

**Display the List of Actors with Actor id and their names**

actors_df

**Normalizing the top movie actors names has been removed and actor id has been included, so that the data frame has been normalized and movie key and actor id has been stored**

top_mov_actors_rev_df = pd.merge(top_mov_actors_df, actors_df,
         left_on='actor',
         right_on='Act_name',
         how='left')

# Drop columns multiple column 
top_mov_actors_rev_df = top_mov_actors_rev_df.drop(['actor', 'Act_name'], axis=1)

top_mov_actors_rev_df

**Storing the Dataframe consoldated data has been stored into the table by creating a database and tables**

#!pip install pandas numpy lets-plot numerize tqdm sqlalchemy jupysql

import os
import sqlite3
import numpy as np
import pandas as pd

from tqdm.notebook import tqdm, trange
tqdm.pandas()

**Database Creation**

DATA_FOLDER = os.path.join('../data/clean')

# Create a SQLite database in the ../data/ directory if it does not already exist
conn = sqlite3.connect(os.path.join(DATA_FOLDER, './supermovies.db'))

**Table Creation**

conn.execute('''
CREATE TABLE IF NOT EXISTS Movie_Actors
(
    Act_id INTEGER,
    Act_name TEXT
)
''')

conn.execute('''
DELETE FROM Movie_Actors
''')             

conn.execute('''
CREATE TABLE IF NOT EXISTS top_movies (
    rank TEXT,
    title TEXT,
    rel_year INT,
    rating INT,
    movie_key TEXT
)
''')

conn.execute('''
DELETE FROM top_movies
''')  

conn.execute('''
CREATE TABLE IF NOT EXISTS top_mov_actors (
    movie_key TEXT,
    Act_id INTEGER
)
''')

conn.execute('''
DELETE FROM top_mov_actors
''')  

**Data has been loaded into the corresponding tables**

actors_df.to_sql('Movie_Actors', conn, if_exists='append', index=False)
top_movies_df.to_sql('top_movies', conn, if_exists='append', index=False)
top_mov_actors_rev_df.to_sql('top_mov_actors', conn, if_exists='append', index=False)


**Checking the data from the database**

%sql sqlite:///../data/clean/supermovies.db --alias supermovies


%sql supermovies

%%sql

SELECT * FROM movie_actors;

%%sql

SELECT * FROM top_mov_actors;


%%sql

SELECT * FROM top_movies;

%%sql

select rel_year, count(rel_year)
from top_movies
group by rel_year
