==> notebooks/NB03-EDA-and-Dataviz.ipynb <==
**NB03-Data-Analysis**


%load_ext sql

import os
import sqlite3
import numpy as np
import pandas as pd

from lets_plot import *
LetsPlot.setup_html()

from tqdm.notebook import tqdm, trange
tqdm.pandas()

DATA_FOLDER = os.path.join('../data/clean')

# Create a SQLite database in the ../data/ directory if it does not already exist
conn = sqlite3.connect(os.path.join(DATA_FOLDER, './supermovies.db'))

%sql sqlite:///../data/clean/supermovies.db --alias supermovies

%sql supermovies

DATA_FOLDER = os.path.join('../data/clean')

# Create a SQLite database in the ../data/ directory if it does not already exist
conn = sqlite3.connect(os.path.join(DATA_FOLDER, './supermovies.db'))

%%sql --alias supermovies year_wise_no_of_mov_df <<

select rel_year, count(rel_year) "Number of Movies"
from top_movies
group by rel_year

type(year_wise_no_of_mov_df)

year_wise_no_of_mov_df = year_wise_no_of_mov_df.DataFrame() # Convert to pandas DataFrame

year_wise_no_of_mov_df

import numpy as np
from lets_plot import *
LetsPlot.setup_html()
#np.random.seed(42)

#ggplot(my_df, aes(x='rel_year', y='Number of Movies')) + geom_point()
ggplot(year_wise_no_of_mov_df) + \
    geom_point(aes(x='rel_year', y='Number of Movies', color='Number of Movies')) + \
        ggtitle('Year wise number of movies released of the Top 97 Movies') + \
        xlab('Release Year') + \
        ylab('Number of Movies')


#ggplot() + geom_point(aes(x='rel_year', y='Number of Movies'), data=data)

ggplot() + \
    geom_bar(aes(x='rel_year', y='Number of Movies', fill='rel_year'), data=year_wise_no_of_mov_df, \
             stat='identity', show_legend=False) + \
             ggtitle('Year wise number of movies released of the Top 97 Movies') + \
        xlab('Release Year') + \
        ylab('Number of Movies')+\
    scale_fill_discrete()

(
    ggplot(year_wise_no_of_mov_df, aes("rel_year", "Number of Movies")) +
    geom_line(size=2)+
    ggtitle('Year wise number of movies released of the Top 97 Movies') + 
    xlab('Release Year') + 
    ylab('Number of Movies')
)

(
   ggplot(year_wise_no_of_mov_df, aes("rel_year", "Number of Movies")) +
    geom_path(size=1) +
    geom_point(size=5) +
    ggtitle('Year wise number of movies released of the Top 97 Movies') + 
    xlab('Release Year') + 
    ylab('Number of Movies')
)

%%sql --alias supermovies actor_wise_no_of_mov_df <<

SELECT ma.act_id, act_name, count(title)"Number of Movies"
  FROM top_movies tm
  left join top_mov_actors tma
  on tm.movie_key = tma.movie_key
  left join movie_actors ma 
  on tma.act_id = ma.act_id
  group by ma.act_name
  having count(tm.title) > 2


# actor_wise_no_of_mov_df = actor_wise_no_of_mov_df.DataFrame() 
print(type(actor_wise_no_of_mov_df))

actor_wise_no_of_mov_df = actor_wise_no_of_mov_df.DataFrame()

actor_wise_no_of_mov_df.info()

(ggplot(data=actor_wise_no_of_mov_df, mapping=aes(x='Act_name', y='Number of Movies', fill='Act_name')) + 
    geom_bar( stat='identity', show_legend=False) +
    xlab('Actor Name') + 
    ylab('Number of Movies') +
    scale_fill_discrete()+
    ggtitle('Top 97 Movies', subtitle='Most Frequent Actors')
    )


(ggplot(data=actor_wise_no_of_mov_df, mapping=aes(x='Act_name', y='Number of Movies', fill='Act_name')) + 
    geom_line( stat='identity', show_legend=False, size=2) +
    xlab('Actor Name') + 
    ylab('Number of Movies') +
    scale_fill_discrete()+
    geom_point(size=5, show_legend=False)+
    ylim(1.8,3.2)+
    ggtitle('Top 97 Movies', subtitle='More than one movies acted by the Actors in the top 97 movies')
    )
